Pre-Server Synchronisation Commit
This commit preserves to a great (but not working) extent the old
LoLTeamChecker app from LoLStatsAnal. Changes will be made so this works only
calls to http://mahennads.pythonanywhere.com/...
A pythonanywhere helper class will be added, data parsers will be modified,
core code (in the MVC) will be changed for the new classes/methods, and
eventually it will be packaged with distools/pip, and maybe py2exe.